# Implement left-to-right decoding