--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "crockford"
-version = "1.0.0-beta.2"
+version = "1.0.0-beta.3"
 authors = ["J/A <archer884@gmail.com>"]
 license = "MIT/Apache-2.0"
 documentation = "https://docs.rs/crockford"